mock for type cache moved into separate file
