+++ dao/mock_daos.go
@@ -949,20 +949,3 @@ func (mad MockAuthenticationDao) ListIdsForResource(resourceType string, resourc
 func (m MockAuthenticationDao) BulkDelete(authentications []m.Authentication) ([]m.Authentication, error) {
 	return authentications, nil
 }
-
-func PopulateMockStaticTypeCache() error {
-	tc := typeCache{}
-	tc.sourceTypes = make(map[string]int64)
-	tc.applicationTypes = make(map[string]int64)
-
-	for _, st := range fixtures.TestSourceTypeData {
-		tc.sourceTypes[st.Name] = st.Id
-	}
-
-	for _, at := range fixtures.TestApplicationTypeData {
-		tc.applicationTypes[at.Name] = at.Id
-	}
-
-	Static = tc
-	return nil
-}
